refactor: SuccessPage ㅇ
